feat: 重构chain_info服务
